MB-67520: Fix dataloss warnings and make them compile errors
For Clang (and includes AppleClang) fix warnings produced by
-Wshorten-64-to-32 -Wimplicit-int-conversion and then leave
those warnings as errors.

Reviewed-on: https://review.couchbase.org/c/platform/+/230427
Tested-by: Build Bot <build@couchbase.com>

+++ cbcrypto/common.cc
@@ -128,8 +128,8 @@ OpenSslError::OpenSslError(const char* callingFunction,
 }
 
 void randomBytes(gsl::span<char> buf) {
-    if (RAND_bytes(reinterpret_cast<unsigned char*>(buf.data()), buf.size()) !=
-        1) {
+    if (RAND_bytes(reinterpret_cast<unsigned char*>(buf.data()),
+                   gsl::narrow_cast<int>(buf.size())) != 1) {
         throw OpenSslError::get("cb::crypto::randomBytes", "RAND_bytes");
     }
 }
